refactor: PlantRepository 통합 및 중복 리포지토리 제거
