+++ core/data/src/main/java/com/simuel/sunflower/core/data/di/RepositoryModule.kt
@@ -1,10 +1,8 @@
 package com.simuel.sunflower.core.data.di
 
 import com.simuel.sunflower.core.data.repository.GardenRepositoryImpl
-import com.simuel.sunflower.core.data.repository.PlantDetailRepositoryImpl
 import com.simuel.sunflower.core.data.repository.PlantRepositoryImpl
 import com.simuel.sunflower.core.domain.repository.GardenRepository
-import com.simuel.sunflower.core.domain.repository.PlantDetailRepository
 import com.simuel.sunflower.core.domain.repository.PlantRepository
 import dagger.Binds
 import dagger.Module
@@ -19,7 +17,4 @@ internal abstract class RepositoryModule {
 
     @Binds
     abstract fun bindGardenRepository(gardenRepository: GardenRepositoryImpl): GardenRepository
-
-    @Binds
-    abstract fun bindPlantDetailRepository(plantDetailRepository: PlantDetailRepositoryImpl): PlantDetailRepository
 }
