--- core/domain/src/main/java/com/simuel/sunflower/core/domain/repository/PlantDetailRepository.kt
@@ -1,7 +1,0 @@
-package com.simuel.sunflower.core.domain.repository
-
-import com.simuel.sunflower.core.domain.model.Plant
-
-interface PlantDetailRepository {
-    fun findPlantById(plantId: String): Plant
-}
