+++ feature/plantdetail/src/main/java/com/simuel/sunflower/feature/plantdetail/PlantDetailViewModel.kt
@@ -2,7 +2,7 @@ package com.simuel.sunflower.feature.plantdetail
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.simuel.sunflower.core.domain.repository.PlantDetailRepository
+import com.simuel.sunflower.core.domain.repository.PlantRepository
 import com.simuel.sunflower.feature.plantdetail.model.PlantDetailUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -13,7 +13,7 @@ import javax.inject.Inject
 
 @HiltViewModel
 class PlantDetailViewModel @Inject constructor(
-    private val repository: PlantDetailRepository
+    private val repository: PlantRepository
 ) : ViewModel() {
     private val _uiState = MutableStateFlow<PlantDetailUiState>(PlantDetailUiState.Loading)
     val uiState: StateFlow<PlantDetailUiState> = _uiState.asStateFlow()
